feat: navchart
