+++ src/pages/past-performance/components/copyright.tsx
@@ -1,10 +1,11 @@
 import React from 'react'
+import { CopyRightProps } from '../type';
 
-function Copyright() {
+function Copyright({ year }: CopyRightProps) {
     return (
         <div>
             <div className='flex justify-center'>
-                <p className='text-[14px] poppins-regular pt-4 pb-2'><span className='text-orange-400'>Copyright ©</span> 2024 Keloola by Thrive</p>
+                <p className='text-[14px] poppins-regular pt-4 pb-2'><span className='text-orange-400'>Copyright ©</span> {year} Keloola by Thrive</p>
             </div>
         </div>
     )
