+++ src/pages/past-performance/type.ts
@@ -7,4 +7,20 @@ export interface HeaderProps {
 export interface TitleProps {
     title: string;
     description: ReactNode; 
+    svg:ReactNode
+}
+
+export interface ItemProps {
+    id: number,
+    tagId: string;
+    title: string;
+    status: boolean;
+}
+
+export interface HoverState {
+    state: number | null;
+}
+
+export interface CopyRightProps {
+    year: number;
 }
